+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ## 0.2.0
 
+- Add `nullPath` for auxiliary types.
 - Add `toList :: ASTPath -> [String]`.
 
 ## 0.1.0